pyproject.toml
[project]
name = "streamit-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.27",
    "fastapi>=0.116.1",
    "langchain-google-genai>=2.1.9",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-postgres>=2.0.23",
    "mcp>=1.13.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "psycopg[binary,pool]>=3.2.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = [
    "-sv",
]
asyncio_default_fixture_loop_scope = "function"
